add iniFile class file
