+++ DataConverter/Converter.cs
@@ -8,7 +8,6 @@ using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 using System.IO;
-using System.Runtime.InteropServices;
 using System.Windows.Forms;
 using System.Security.Cryptography;
 
@@ -705,29 +704,6 @@ namespace DataConverter
         }
     }
 
-    public class InitFile
-    {
-        public static string IniPath = Path.Combine(Application.StartupPath, "config.ini");
 
-        [DllImport("kernel32")]
-        private static extern long WritePrivateProfileString(string section, string key, string val, string filePath);
-        [DllImport("kernel32")]
-        private static extern int GetPrivateProfileString(string section, string key, string def, StringBuilder retVal, int size, string filePath);
-
-        public InitFile(string path)
-        {
-
-        }
-
-        public static string ReadValue(string Section, string Key)
-        {
-            StringBuilder temp = new StringBuilder(255);
-            int i = GetPrivateProfileString(Section, Key, "", temp, 255, IniPath);
-            return temp.ToString();
-        }
-
-
-
-    }
 
 }
